requirements.txt
torch
torchvision
pillow
matplotlib
numpy
opencv-python
gradio
